+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/perks/BigPerks.java
@@ -390,3 +390,5 @@ public class BigPerks implements ISlashRegistryInit {
         );
     }
 }
+
+