+++ src/main/java/com/robertx22/age_of_exile/loot/blueprints/ItemBlueprint.java
@@ -1,7 +1,7 @@
 package com.robertx22.age_of_exile.loot.blueprints;
 
-import com.robertx22.age_of_exile.database.registry.RarityRegistryContainer;
 import com.robertx22.age_of_exile.database.registry.Database;
+import com.robertx22.age_of_exile.database.registry.RarityRegistryContainer;
 import com.robertx22.age_of_exile.loot.LootInfo;
 import com.robertx22.age_of_exile.loot.blueprints.bases.LevelPart;
 import com.robertx22.age_of_exile.loot.blueprints.bases.TierPart;
@@ -40,7 +40,7 @@ public abstract class ItemBlueprint {
         this.onConstruct();
     }
 
-    protected List<IStackAction> actionsAfterGeneration = new ArrayList<>();
+    public List<IStackAction> actionsAfterGeneration = new ArrayList<>();
 
     boolean itemWasGenerated = false;